*** empty log message ***

+++ trunk/ace/ReactorEx.cpp
@@ -179,7 +179,7 @@ ACE_ReactorEx::handle_events (ACE_Time_Value *how_long,
 // handles_[active_handles_] using WaitForMultipleObjects to poll
 // through our handle set looking for active handles.
 int
-ACE_ReactorEx::dispatch_all (int index, int wait_all)
+ACE_ReactorEx::dispatch_all (size_t index, int wait_all)
 {
   while (1)
     {
